fix: move App.tsx
